feat: Implement pages for HackUTD VIII site
The index, faq, schedule, speakers, and sponsors were created and updated the
homepage. Font files, style sheet, and photos were added to style the homepage.
SiteHeader component was added to replace repeated HTML code within the
aforementioned pages. The faq, schedule, speakers, and sponsors pages are stub
pages and will require future changes to fulfill the design of the website and
the index page will require minor style changes for desktop and the
implementation of the mobile version.