feat: Add dropdown feature to faq page
The faq page now uses components to have a dropdown from the question to answer.
The faq page is also completely styled for desktop through the globals.css file.
The dropdown works through states and when user clicks on the question. Then the
answer which is another component will drop down.