feat: Add mobile navigation
All pages on the HackUTD site now has mobile navigation. Utilizing components
and states when a user clicks on the logo then navigation opens. The navigation
is partly reflects the design in the figma besides icons. In addition, the
components mobile header and mobile links are added to help organization of
navigation.